+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 💻 Changes & Update
 
+-   Create Header & Footer Components
 -   Create Slideshow component to show list of recommended komiks
 -   Create KontenX component to show list of komiks horizontaly
 -   Create GenreMenu component to show list of genres available